Разбивка Engine.py на Low и High Objects
Разбил модуль Engine на две составляющие.

Low - Coordinates, Point, Vector, Vector space
High - Camera, Parameters, Object
