+++ Code/II sem/Analityc geometry/Projects/Console3D/config.cfg
@@ -1,3 +1,3 @@
 [SCREEN_PARAM]
-height = 40
-width = 70
+height = 800
+width = 600
